requirements.txt
fastapi
uvicorn
mediapipe
numpy
opencv-python
pillow
python-multipart